Refactor context setup
