+++ src/dao_frontend/src/main.jsx
@@ -2,13 +2,16 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
 import { ActorProvider } from './context/ActorContext'
+import { AuthProvider } from './context/AuthContext'
 import './index.css'
 import './app.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <ActorProvider>
-      <App />
-    </ActorProvider>
+    <AuthProvider>
+      <ActorProvider>
+        <App />
+      </ActorProvider>
+    </AuthProvider>
   </React.StrictMode>,
 )
